backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
deprecation==2.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
frozenlist==1.4.1
-e .
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
postgrest==0.16.11
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
realtime==2.0.5
redis==5.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.6
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.31.0
wcwidth==0.2.13
websockets==13.1
yarl==1.13.1
zipp==3.20.2
